# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Backend API
  version: "1.0.0"
  description: |
    REST API for the Vite + React PWA.

    ## Response envelope

    Every response uses one of two shapes.

    Success:
    ```json
    { "success": true, "data": { }, "meta": { } }
    ```
    `meta` is present only on paginated collections.

    Error:
    ```json
    { "success": false, "message": "...", "code": "VALIDATION_ERROR", "errors": [], "request_id": "..." }
    ```
    `errors` carries per-field messages for validation failures. `request_id`
    matches the `X-Request-ID` response header and the server logs.

    ## Authentication

    Send the access token as `Authorization: Bearer <token>`. Access tokens are
    short lived (15 minutes by default); use `POST /auth/refresh` to rotate the
    refresh token into a new pair. Refresh tokens are single use: presenting one
    twice is treated as theft and revokes every session for the account.
  license:
    name: MIT
  contact:
    name: API support
    email: api@example.com
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
  - url: https://api.example.com/api/v1
    description: Production
...
security:
  - BearerAuth: []
paths:
...
  /users:
    get:
      tags: [users]
      summary: List users
      description: Administrators only. Supports pagination, sorting, search and filters.
      operationId: listUsers
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
          example: 1
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
          example: 20
        - name: sort
          in: query
          schema:
            type: string
            enum: [created_at, updated_at, email, full_name]
            default: created_at
        - name: order
          in: query
          schema: { type: string, enum: [asc, desc], default: desc }
        - name: search
          in: query
          description: Case-insensitive match against email or full name.
          schema: { type: string, maxLength: 128 }
          example: jane
        - name: role
          in: query
          schema: { type: string, enum: [user, admin] }
        - name: is_active
          in: query
          schema: { type: boolean }
      responses:
        "200":
          description: A page of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListEnvelope"
              example:
                success: true
                data:
                  - id: 3f0d1c1e-8f4e-4a4b-9c2a-2f9a1b7d9e11
                    email: jane@example.com
                    full_name: Jane Doe
                    role: user
                    avatar_url: null
                    is_active: true
                    created_at: "2026-07-30T09:12:33Z"
                    updated_at: "2026-07-30T09:12:33Z"
                meta:
                  page: 1
                  limit: 20
                  total: 1
                  total_pages: 1
                  has_next: false
                  has_prev: false
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationError"

    post:
      tags: [users]
      summary: Create a user
      description: Administrators only. Unlike registration, the role may be set explicitly.
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
            example:
              email: newadmin@example.com
              password: Sup3rSecretPassword
              full_name: New Admin
              role: admin
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserEnvelope"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
  /users/me:
    get:
      tags: [users]
      summary: Get the current user
      operationId: getCurrentUser
      responses:
        "200":
          description: The authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
    patch:
      tags: [users]
      summary: Update the current user
      description: |
        Partial update. `role` and `is_active` are rejected for non-administrators.
      operationId: updateCurrentUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
            example:
              full_name: Jane R. Doe
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserEnvelope"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationError"
  /users/me/avatar:
    post:
      tags: [users]
      summary: Upload an avatar
      description: |
        Multipart upload. Accepts JPEG, PNG, WebP and GIF up to
        `STORAGE_MAX_UPLOAD_BYTES`. The previous avatar is deleted afterwards.
      operationId: uploadAvatar
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [avatar]
              properties:
                avatar:
                  type: string
                  format: binary
      responses:
        "200":
          description: Updated user with the new avatar URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserEnvelope"
        "413":
          description: The file is too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "415":
          description: Unsupported image type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /users/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
        example: 3f0d1c1e-8f4e-4a4b-9c2a-2f9a1b7d9e11
    get:
      tags: [users]
      summary: Get a user
      description: Administrators can read any account; everyone else only their own.
      operationId: getUser
      responses:
        "200":
          description: The user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserEnvelope"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [users]
      summary: Update a user
      description: |
        Changing `role` or `is_active` requires the admin role and revokes the
        target account's sessions.
      operationId: updateUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
            example:
              role: admin
              is_active: true
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserEnvelope"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [users]
      summary: Delete a user
      description: Administrators can delete any account except their own.
      operationId: deleteUser
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageEnvelope"
              example:
                success: true
                data:
                  message: User deleted
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Access token from /auth/login. Header: `Authorization: Bearer <token>`."
  schemas:
    User:
      type: object
      required: [id, email, full_name, role, is_active, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        full_name: { type: string }
        role: { type: string, enum: [user, admin] }
        avatar_url: { type: string, nullable: true }
        is_active: { type: boolean }
        email_verified_at: { type: string, format: date-time, nullable: true }
        last_login_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    PageMeta:
      type: object
      properties:
        page: { type: integer, example: 1 }
        limit: { type: integer, example: 20 }
        total: { type: integer, format: int64, example: 42 }
        total_pages: { type: integer, example: 3 }
        has_next: { type: boolean }
        has_prev: { type: boolean }
    FieldError:
      type: object
      properties:
        field: { type: string, example: email }
        message: { type: string, example: Must be a valid email address }
    ErrorEnvelope:
      type: object
      required: [success, message]
      properties:
        success: { type: boolean, example: false }
        message: { type: string }
        code:
          type: string
          enum:
            - BAD_REQUEST
            - VALIDATION_ERROR
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - CONFLICT
            - TOO_MANY_REQUESTS
            - PAYLOAD_TOO_LARGE
            - UNSUPPORTED_MEDIA_TYPE
            - INTERNAL_ERROR
            - SERVICE_UNAVAILABLE
        errors:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"
        request_id: { type: string }
    MessageEnvelope:
      type: object
      properties:
        success: { type: boolean, example: true }
        data:
          type: object
          properties:
            message: { type: string }
...
    UserEnvelope:
      type: object
      properties:
        success: { type: boolean, example: true }
        data:
          $ref: "#/components/schemas/User"
    UserListEnvelope:
      type: object
      properties:
        success: { type: boolean, example: true }
        data:
          type: array
          items:
            $ref: "#/components/schemas/User"
        meta:
          $ref: "#/components/schemas/PageMeta"
...
    CreateUserRequest:
      type: object
      required: [email, password, full_name]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 12, maxLength: 72 }
        full_name: { type: string, minLength: 2, maxLength: 120 }
        role: { type: string, enum: [user, admin], default: user }
    UpdateUserRequest:
      type: object
      description: Partial update; omitted fields are left unchanged.
      properties:
        full_name: { type: string, minLength: 2, maxLength: 120 }
        role:
          type: string
          enum: [user, admin]
          description: Administrators only.
        is_active:
          type: boolean
          description: Administrators only.
  responses:
    Unauthorized:
      description: Authentication is required or the token is no longer valid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            success: false
            message: Authentication is required
            code: UNAUTHORIZED
    Forbidden:
      description: The caller lacks permission
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            success: false
            message: You do not have permission to perform this action
            code: FORBIDDEN
    NotFound:
      description: The resource does not exist
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            success: false
            message: User not found
            code: NOT_FOUND
    Conflict:
      description: The resource already exists
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            success: false
            message: An account with this email already exists
            code: CONFLICT
    ValidationError:
      description: One or more fields are invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            success: false
            message: The request contains invalid fields
            code: VALIDATION_ERROR
            errors:
              - field: email
                message: Must be a valid email address
              - field: password
                message: Must be at least 12 characters
            request_id: 0f2b0e2c-3a1d-4c9a-9e21-6b7c8d9e0f11
...